new ixa syntax for validator in define_global_properties
